requirements.txt
streamlit
torch
torchvision
ultralytics
Pillow
matplotlib
